requirements.txt
discord.py>=2.3.2
coc.py
python-dotenv
aiohttp
